Add althea-cron-jobs
This adds a package for periodic task scheduling, it's really just
a dummy package to enable cron on boot and provide a cron file
and update script. The big advantage of having it as a package
is that it can update itself (in theory). By default it runs
every day at 4:30am and checks for updates to Althea packages
before rebooting if any are found.