No more startup delay, pipe to null
Turns out that if you have a script that lets stdout attach to your tty
disconnecting kills the process. This prevents the checkup script from
working properly as it always kills Rita as it tries to restart it.

The 30 second delay seems to only make this worse, so it's also gone.