Althea cron jobs should use /etc/init.d and install_bin
So there are two problems with Althea cron jobs, mainly that it was
installing into rc.d which is not supposed to contain the actual scripts
and second that it was using cp for an init script, resulting in issues with
that file being not exacutable. Somehow this was caught and fixed during the
firmware build process but not during the package install process.

This commit fixes both issues and should make the cron-jobs package
far more robust.