Update default Babeld config
The previous contents of this file do not work with Althea, but this
problem was hidden by the fact that packages don't overwrite locally
created or locally changed files by default. Meaning the Babeld config
templated by the firmware builder would remain in place and not be overwritten
with this incorrect file.

Now that this is updated to be correct we can stop templating the file over in
Althea Firmware since it's now static anyways