Fix commit hashes for packages
After a loony-tunes esque series of build failures I've decided that this
is the best way to go about things.

While building off of master allows for quick prototyping and makes it easier
for people to get all new builds to test it also makes it very difficult to
impossible to roll back time and make minor changes to actual release builds.

The latter is probably more important than the former, especially when under
a time crunch so this is probably the best way to go forward.